Migrate operation form to it own files
